login form separated into separate component
